Revert "improved-performance"
