+++ app/(default)/page.tsx
@@ -1,4 +1,4 @@
-import Features from "@/components/features/Features";
+import Features from "@/components/features";
 import Hero from "@/components/hero";
 import Showcase from "@/components/showcase";
 import { constructMetadata } from "@/lib/utils";
