+++ app/layout.tsx
@@ -9,6 +9,7 @@ import dynamic from "next/dynamic";
 import { Analytics } from "@vercel/analytics/react";
 import { SpeedInsights } from "@vercel/speed-insights/next";
 import { ThemeProvider } from "next-themes";
+import { ReactLenis } from "@/utils/lenis";
 
 const PostHogPageView = dynamic(() => import("./PostHogPageView"), {
   ssr: false,
@@ -22,24 +23,28 @@ export default function RootLayout({
   return (
     <html lang="en" className={`${GeistSans.variable} ${GeistMono.variable}`}>
       <PHProvider>
-        <body className={`bg-background font-sans tracking-tight antialiased`}>
-          <ThemeProvider
-            attribute="class"
-            enableSystem
-            disableTransitionOnChange
-            enableColorScheme
+        <ReactLenis root>
+          <body
+            className={`bg-background font-sans tracking-tight antialiased`}
           >
-            <PostHogPageView />
-            <div className="flex min-h-screen flex-col overflow-hidden">
-              <Header />
-              {children}
-              <Toaster position="bottom-right" richColors />
-              <Analytics />
-              <SpeedInsights />
-            </div>
-          </ThemeProvider>
-          <Analytics />
-        </body>
+            <ThemeProvider
+              attribute="class"
+              enableSystem
+              disableTransitionOnChange
+              enableColorScheme
+            >
+              <PostHogPageView />
+              <div className="flex min-h-screen flex-col overflow-hidden">
+                <Header />
+                {children}
+                <Toaster position="bottom-right" richColors />
+                <Analytics />
+                <SpeedInsights />
+              </div>
+            </ThemeProvider>
+            <Analytics />
+          </body>
+        </ReactLenis>
       </PHProvider>
     </html>
   );
