+++ components/ui/expandable-cards.tsx
@@ -12,10 +12,9 @@ import {
 
 type Card = {
   id: number;
-  title: string;
-  company: {
-    name: string;
-    url: string;
+  title: {
+    desktop: JSX.Element;
+    mobile: JSX.Element;
   };
   description: string | JSX.Element;
   icon: React.ComponentType<{ className?: string; color?: string }>;
@@ -29,10 +28,34 @@ type Card = {
 const CARDS: Card[] = [
   {
     id: 1,
-    title: "Ship with",
-    company: {
-      name: "aider",
-      url: "https://aider.chat/",
+    title: {
+      desktop: (
+        <>
+          Ship with
+          <br />
+          <a
+            href="https://aider.chat/"
+            target="_blank"
+            rel="noopener noreferrer"
+            className="hover:underline"
+          >
+            aider
+          </a>
+        </>
+      ),
+      mobile: (
+        <>
+          Ship with{" "}
+          <a
+            href="https://aider.chat/"
+            target="_blank"
+            rel="noopener noreferrer"
+            className="hover:underline"
+          >
+            aider
+          </a>
+        </>
+      ),
     },
     description:
       "Aider is a popular CLI tool that generates code for you. Ask for a new feature, a refactor, or to fix a bug. Integrated within PearAI Creator, it will make and apply the changes to your files automatically.",
@@ -45,10 +68,34 @@ const CARDS: Card[] = [
   },
   {
     id: 2,
-    title: "Ship with",
-    company: {
-      name: "Supermaven",
-      url: "https://supermaven.com/",
+    title: {
+      desktop: (
+        <>
+          Predict with
+          <br />
+          <a
+            href="https://supermaven.com/"
+            target="_blank"
+            rel="noopener noreferrer"
+            className="hover:underline"
+          >
+            Supermaven
+          </a>
+        </>
+      ),
+      mobile: (
+        <>
+          Predict with{" "}
+          <a
+            href="https://supermaven.com/"
+            target="_blank"
+            rel="noopener noreferrer"
+            className="hover:underline"
+          >
+            Supermaven
+          </a>
+        </>
+      ),
     },
     description:
       "Supermaven is currently the fastest code autocomplete tool available. Supermaven comes independently pre-packaged and recommended within PearAI.",
@@ -61,10 +108,35 @@ const CARDS: Card[] = [
   },
   {
     id: 3,
-    title: "Chat & edit with",
-    company: {
-      name: "Continue",
-      url: "https://www.continue.dev/",
+    title: {
+      desktop: (
+        <>
+          Chat & edit
+          <br />
+          with{" "}
+          <a
+            href="https://www.continue.dev/"
+            target="_blank"
+            rel="noopener noreferrer"
+            className="hover:underline"
+          >
+            Continue
+          </a>
+        </>
+      ),
+      mobile: (
+        <>
+          Chat & edit with{" "}
+          <a
+            href="https://www.continue.dev/"
+            target="_blank"
+            rel="noopener noreferrer"
+            className="hover:underline"
+          >
+            Continue
+          </a>
+        </>
+      ),
     },
     description: (
       <>
@@ -83,10 +155,34 @@ const CARDS: Card[] = [
   },
   {
     id: 4,
-    title: "Remember with",
-    company: {
-      name: "Mem0",
-      url: "https://mem0.ai/",
+    title: {
+      desktop: (
+        <>
+          Remember
+          <br />
+          <a
+            href="https://mem0.ai/"
+            target="_blank"
+            rel="noopener noreferrer"
+            className="hover:underline"
+          >
+            with Mem0
+          </a>
+        </>
+      ),
+      mobile: (
+        <>
+          Remember{" "}
+          <a
+            href="https://mem0.ai/"
+            target="_blank"
+            rel="noopener noreferrer"
+            className="hover:underline"
+          >
+            with Mem0
+          </a>
+        </>
+      ),
     },
     description:
       "Mem0 is a self-improving memory layer for LLM applications. It is integrated into PearAI to make coding with AI an enhanced, personalized experience.",
@@ -99,10 +195,34 @@ const CARDS: Card[] = [
   },
   {
     id: 5,
-    title: "Search with",
-    company: {
-      name: "Perplexity",
-      url: "https://www.perplexity.ai/",
+    title: {
+      desktop: (
+        <>
+          Search with
+          <br />
+          <a
+            href="https://www.perplexity.ai/"
+            target="_blank"
+            rel="noopener noreferrer"
+            className="hover:underline"
+          >
+            Perplexity
+          </a>
+        </>
+      ),
+      mobile: (
+        <>
+          Search with{" "}
+          <a
+            href="https://www.perplexity.ai/"
+            target="_blank"
+            rel="noopener noreferrer"
+            className="hover:underline"
+          >
+            Perplexity
+          </a>
+        </>
+      ),
     },
     description:
       "Perplexity is the leading AI search engine. Integrated within PearAI Search, you can seamlessly add web content, like up-to-date documentation, which vanilla LLMs often lack, directly into your requests.",
@@ -196,9 +316,7 @@ export default function ExpandableCards() {
                           }}
                           className="whitespace-pre-line text-xl font-[550] leading-6"
                         >
-                          {card.title}
-                          <br />
-                          <span>{card.company.name}</span>
+                          {card.title.desktop}
                           <span className="opacity-50">*</span>
                         </h1>
                       </motion.div>
@@ -210,23 +328,14 @@ export default function ExpandableCards() {
                       transition={{ delay: ANIMATION_DELAY }}
                       className="flex h-full flex-col"
                     >
-                      <div
+                      <h4
                         style={{
                           color: card.titleColor,
                         }}
                         className="mb-3 text-xl font-[550] leading-6"
                       >
-                        {card.title}{" "}
-                        <a
-                          href={card.company.url}
-                          target="_blank"
-                          rel="noopener noreferrer"
-                          className="hover:underline"
-                        >
-                          {" "}
-                          {card.company.name}
-                        </a>
-                      </div>
+                        {card.title.mobile}
+                      </h4>
                       <p
                         style={{
                           color: card.descriptionColor,
@@ -277,19 +386,7 @@ export default function ExpandableCards() {
                         style={{ color: card.titleColor }}
                         className="text-base font-semibold sm:text-lg"
                       >
-                        {card.title}{" "}
-                        {expandedId === card.id ? (
-                          <a
-                            href={card.company.url}
-                            target="_blank"
-                            rel="noopener noreferrer"
-                            className="hover:underline"
-                          >
-                            {card.company.name}
-                          </a>
-                        ) : (
-                          <span>{card.company.name}</span>
-                        )}
+                        {card.title.mobile}
                       </div>
                     </div>
                     <div
